fix browser compat, take 2
