+++ rate_limit/memory_store.ts
@@ -1,7 +1,10 @@
 // Copyright 2018-2026 the Deno authors. MIT license.
 
-import type { RateLimitResult } from "./rate_limiter.ts";
-import type { AlgorithmOptions, RateLimitStore } from "./store_types.ts";
+import type {
+  AlgorithmOptions,
+  RateLimitResult,
+  RateLimitStore,
+} from "./store_types.ts";
 import {
   assertNonNegativeInteger,
   assertPositiveFinite,
